moving dateValidation
